Separating factory and client
